CRAYSAT-1568: Allow empty files to pass license check
This commit changes the license check so that empty files are always
allowed by the license checker. This commit also adds a unit test
for this behavior.

Add `if __name__ == '__main__'` guard around main code in the
test script, so that unit tests can be discovered and run with
a test runner.